Addressing offline feedback
