+++ src/Microsoft.AspNetCore.Server.HttpSys/AsyncAcceptContext.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.HttpSys.Internal;
@@ -16,12 +17,14 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         private TaskCompletionSource<RequestContext> _tcs;
         private HttpSysListener _server;
         private NativeRequestContext _nativeRequestContext;
+        private const int DefaultBufferSize = 4096;
+        private const int AlignmentPadding = 8;
 
         internal AsyncAcceptContext(HttpSysListener server)
         {
             _server = server;
             _tcs = new TaskCompletionSource<RequestContext>();
-            _nativeRequestContext = HttpApi.AllocateNativeRequest(this);
+            _nativeRequestContext = AllocateNativeRequest(this);
         }
 
         internal Task<RequestContext> Task
@@ -80,13 +83,13 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                         finally
                         {
                                 // The request has been handed to the user, which means this code can't reuse the blob.  Reset it here.
-                            asyncResult._nativeRequestContext = complete ? null : HttpApi.AllocateNativeRequest(asyncResult, size: asyncResult._nativeRequestContext.Size);
+                            asyncResult._nativeRequestContext = complete ? null : AllocateNativeRequest(asyncResult, size: asyncResult._nativeRequestContext.Size);
                         }
                     }
                     else
                     {
                         //  (uint)backingBuffer.Length - AlignmentPadding
-                        asyncResult._nativeRequestContext = HttpApi.AllocateNativeRequest(asyncResult, numBytes);
+                        asyncResult._nativeRequestContext = AllocateNativeRequest(asyncResult, numBytes, asyncResult._nativeRequestContext.RequestId);
                     }
 
                     // We need to issue a new request, either because auth failed, or because our buffer was too small the first time.
@@ -158,7 +161,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                     // the buffer was not big enough to fit the headers, we need
                     // to read the RequestId returned, allocate a new buffer of the required size
                     //  (uint)backingBuffer.Length - AlignmentPadding
-                    _nativeRequestContext = HttpApi.AllocateNativeRequest(this, bytesTransferred);
+                    _nativeRequestContext = AllocateNativeRequest(this, bytesTransferred);
                     retry = true;
                 }
                 else if (statusCode == UnsafeNclNativeMethods.ErrorCodes.ERROR_SUCCESS
@@ -170,6 +173,37 @@ namespace Microsoft.AspNetCore.Server.HttpSys
             }
             while (retry);
             return statusCode;
+        }
+
+        // Take this out to a static method that returns a native request context. 
+        internal static NativeRequestContext AllocateNativeRequest(AsyncAcceptContext acceptResult, uint? size = null, ulong requestId = 0)
+        {
+            //Debug.Assert(size != 0, "unexpected size");
+
+            // We can't reuse overlapped objects
+            uint newSize = size.HasValue ? size.Value : DefaultBufferSize;
+            var backingBuffer = new byte[newSize + AlignmentPadding];
+
+            var boundHandle = acceptResult.Server.RequestQueue.BoundHandle;
+            var nativeOverlapped = new SafeNativeOverlapped(boundHandle,
+                boundHandle.AllocateNativeOverlapped(IOCallback, acceptResult, backingBuffer));
+
+            var requestAddress = Marshal.UnsafeAddrOfPinnedArrayElement(backingBuffer, 0);
+
+            // TODO:
+            // Apparently the HttpReceiveHttpRequest memory alignment requirements for non - ARM processors
+            // are different than for ARM processors. We have seen 4 - byte - aligned buffers allocated on
+            // virtual x64/x86 machines which were accepted by HttpReceiveHttpRequest without errors. In
+            // these cases the buffer alignment may cause reading values at invalid offset. Setting buffer
+            // alignment to 0 for now.
+            // 
+            // _bufferAlignment = (int)(requestAddress.ToInt64() & 0x07);
+
+            var bufferAlignment = 0;
+
+            var nativeRequest = (HttpApiTypes.HTTP_REQUEST*)(requestAddress + bufferAlignment);
+            // nativeRequest
+            return new NativeRequestContext(nativeOverlapped, bufferAlignment, nativeRequest, backingBuffer, requestId);
         }
 
         public object AsyncState
